Raspberry driver to deploy environments on RPI4 and RPI3B+

+++ worker_exec.py
@@ -1,16 +1,16 @@
 from database.connector import open_session, close_session
-from database.tables import Action, ActionProperty, Node
+from database.tables import Action, ActionProperty, Node, NodeProperty
 from datetime import datetime
 from importlib import import_module
 from lib.config_loader import DATE_FORMAT, load_config
 from sqlalchemy import or_
-import logging, os, time
+import logging, os, subprocess, sys, time
+
 
 # Import the action driver from config_worker.json
-# Import the exec_action function
 action_driver = load_config()["action_driver"]
-py_module = import_module("%s.action_exec" % action_driver)
-exec_action_fct = getattr(py_module, "exec_action_fct")
+# Import the action executor module
+exec_action_mod = import_module("%s.action_exec" % action_driver)
 # Import the PROCESS and STATE_DESC variables
 py_module = import_module("%s.states" % action_driver)
 PROCESS = getattr(py_module, "PROCESS")
@@ -47,6 +47,22 @@ def next_state_move(db_action):
     return True
 
 
+def new_action(db_node, db):
+    act_prop = db.query(ActionProperty
+        ).filter(ActionProperty.node_name == node.name
+        ).filter(or_(ActionProperty.prop_name == "name", ActionProperty.prop_name == "environment")
+        ).all()
+    act = Action()
+    for prop in act_prop:
+        if prop.prop_name == "name":
+            act.name = prop.prop_value
+            act.node_name = node.name
+            act.node_ip = node.ip
+        if prop.prop_name == "environment":
+            act.environment = prop.prop_value
+    return act
+
+
 def init_action_process(db_action, process_name):
     db_action.process = process_name
     db_action.state_idx = None
@@ -65,30 +81,79 @@ def load_reboot_state(db_action):
                 db_action.state_idx = None
             else:
                 db_action.state_idx = idx - 1
-            db_action.reboot_state = None
             next_state_move(action)
         else:
             logging.error("[%s] can not find the process for the '%s' state" % (
                 db_action.node_name, db_action.reboot_state))
 
 
+def load_lost_state(db_node, db):
+    if db_node.lost_state is not None and len(db_node.lost_state) > 0:
+        logging.info("[%s] load the lost state '%s'" % (db_node.name, db_node.lost_state))
+        process_info = db_node.lost_state.split("?!")
+        if len(process_info) == 2:
+            # Create a new action to continue the process
+            act = new_action(db_node, db)
+            act.process = process_info[0]
+            idx = int(process_info[1])
+            if idx == 0:
+                act.state_idx = None
+            else:
+                act.state_idx = idx - 1
+            db.add(act)
+            next_state_move(act)
+            db_node.status = "in_progress"
+            db_node.lost_state = None
+        else:
+            logging.error("[%s] can not find the process for the '%s' state" % (
+                db_node.name, db_node.lost_state))
+
+
 if __name__ == "__main__":
     # This file used by the SystemD service
     STOP_FILE = "tasksstop"
     # Logging configuration
-    logging.basicConfig(filename='info_exec.log', level=logging.INFO)
+    logging.basicConfig(filename='info_exec.log', level=logging.INFO,
+        format='%(asctime)s %(levelname)-8s %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
     # Detect the final states
     final_states = [ "lost" ]
     for state in STATE_DESC:
         if not STATE_DESC[state]["exec"] and not STATE_DESC[state]["post"]:
             final_states.append(state)
     logging.info("### Final states: %s" % final_states)
-    # Try to configure the lost actions
+    # Connection to the database
     db = open_session()
-    lost_actions = db.query(Action).filter(Action.state == "lost").all()
-    for action in lost_actions:
-        logging.info("[%s] lost action rescue" % action.node_name)
-        load_reboot_state(action)
+    # Delete old information about the pimaster
+    pimaster_info = db.query(NodeProperty).filter(NodeProperty.name == "pimaster").all()
+    for info in pimaster_info:
+        db.delete(info)
+    # Register the information about the pimaster (me)
+    cmd = "whoami"
+    process = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
+    my_user = process.stdout.decode("utf-8").strip()
+    cmd = "hostname -I"
+    process = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
+    my_ip = process.stdout.decode("utf-8").strip()
+    if len(my_user) > 0 and len(my_ip) > 0 and len(my_ip.split(".")) == 4:
+        user_record = NodeProperty()
+        user_record.name = "pimaster"
+        user_record.prop_name = "user"
+        user_record.prop_value = my_user
+        db.add(user_record)
+        ip_record = NodeProperty()
+        ip_record.name = "pimaster"
+        ip_record.prop_name = "ip"
+        ip_record.prop_value = my_ip
+        db.add(ip_record)
+        logging.info("pimaster ip: %s, pimaster user: %s" % (my_ip, my_user))
+    else:
+        logging.error("can not get the IP or the user of the pimaster: ip=%s, user=%s" % (my_ip, my_user))
+        sys.exit(13)
+    # Try to configure the lost nodes
+    lost_nodes = db.query(Node).filter(Node.status == "lost").all()
+    for node in lost_nodes:
+        logging.info("[%s] lost node rescue" % node.name)
+        load_lost_state(node, db)
     close_session(db)
     # Analyzing the database
     while not os.path.isfile(STOP_FILE):
@@ -108,24 +173,15 @@ if __name__ == "__main__":
                 # As the Node.name is the primary key, only one node should be selected
                 for n in node:
                     n.status = action.state
+                    if action.state == "lost":
+                        n.lost_state = action.reboot_state
                 # Delete the action
                 db.delete(action)
             # Start actions for the recently configured nodes
             pending_nodes = db.query(Node).filter(Node.status == "ready").all()
             for node in pending_nodes:
-                logging.info("[%s] starts the deploy process" % node.name) 
-                act_prop = db.query(ActionProperty
-                    ).filter(ActionProperty.node_name == node.name
-                    ).filter(or_(ActionProperty.prop_name == "name", ActionProperty.prop_name == "environment")
-                    ).all()
-                act = Action()
-                for prop in act_prop:
-                    if prop.prop_name == "name":
-                        act.name = prop.prop_value
-                        act.node_name = node.name
-                        act.node_ip = node.ip
-                    if prop.prop_name == "environment":
-                        act.environment = prop.prop_value
+                logging.info("[%s] starts the deploy process" % node.name)
+                act = new_action(node, db)
                 init_action_process(act, "deploy")
                 node.status = "in_progress"
                 db.add(act)
@@ -149,9 +205,15 @@ if __name__ == "__main__":
                             state_fct = action.state + "_exec"
                         else:
                             state_fct = action.state + "_post"
-                    logging.info("[%s] enters in '%s' state" % (action.node_name, action.state))
                     # Execute the function associated to the action state
-                    if exec_action_fct(state_fct, action):
+                    action_ret = False
+                    try:
+                        action_ret = getattr(exec_action_mod, state_fct)(action, db)
+                    except:
+                        logging.exception("[%s]" % action.node_name)
+                        sys.exit(42)
+                    if action_ret:
+                        logging.info("[%s] successfully executes '%s'" % (action.node_name, state_fct))
                         # Update the state of the action
                         if state_fct.endswith("_exec") and STATE_DESC[state]["post"]:
                             # Execute the '_post' function
@@ -161,6 +223,7 @@ if __name__ == "__main__":
                             next_state_move(action)
                     else:
                         # The node is not ready, test the reboot timeout
+                        logging.warning("[%s] fails to execute '%s'" % (action.node_name, state_fct))
                         if action.updated_at is None:
                             action.updated_at = datetime.now().strftime(DATE_FORMAT)
                         updated = datetime.strptime(str(action.updated_at), DATE_FORMAT)
@@ -168,15 +231,17 @@ if __name__ == "__main__":
                         action_state = action.state.replace("_exec", "").replace("_post","")
                         reboot_timeout = STATE_DESC[action_state]["before_reboot"]
                         do_lost = True
-                        if reboot_timeout > 0 and action.process != "reboot":
+                        reboot_str = "%s?!%d" % (action.process, action.state_idx)
+                        if reboot_timeout > 0 and \
+                            action.process != "reboot" and action.reboot_state != reboot_str:
+                            do_lost = False
                             if elapsedTime > reboot_timeout:
                                 logging.warning("[%s] hard reboot the node" % action.node_name)
                                 # Remember the last state of the current process
-                                action.reboot_state = "%s?!%d" % (action.process, action.state_idx)
+                                action.reboot_state = reboot_str
                                 init_action_process(action, "reboot")
                             else:
-                                do_lost = False
-                                logging.info("[%s] not ready since %d seconds" %(action.node_name, elapsedTime))
+                                logging.info("[%s] not ready since %d seconds" % (action.node_name, elapsedTime))
                         # The node is not ready, test the lost timeout
                         lost_timeout = STATE_DESC[action_state]["lost"]
                         if do_lost and lost_timeout > 0:
@@ -184,7 +249,7 @@ if __name__ == "__main__":
                                 logging.warning("[%s] is lost. Stop monitoring it!" % action.node_name)
                                 if action.process != "reboot":
                                     # Remember the last state of the current process
-                                    action.reboot_state = "%s?!%d" % (action.process, action.state_idx)
+                                    action.reboot_state = reboot_str
                                 action.state = "lost"
                             else:
                                 logging.info("[%s] not ready since %d seconds" %(action.node_name, elapsedTime))
